Update index.md
Nipping the 'why we organize' section because the blurb at the top of the page -- "Guided by our vision for an inclusive & equitable tech industry, TWC organizes to build worker power through rank & file self-organization and education" -- is clear & compelling, and makes it redundant. Also, without this section, 'who we are' flows nicely into 'who we support'.