Include timezone data for consistency
We use the [TZInfo gem][0] for timezone data. To quote its docs:

> By default, TZInfo will attempt to use TZInfo::Data. If TZInfo::Data
> is not available, then TZInfo will search for a zoneinfo
> directory instead.

We weren't using `TZInfo::Data`, so we were at the mercy of the system's
zoneinfo database.

My system lacked at least one of the timezones we use, `US/Pacific`, so
I got this error when running `jekyll serve` (reformatted slightly):

    .../gems/tzinfo-2.0.6/lib/tzinfo/data_source.rb:321:
    in 'TZInfo::DataSource#validate_timezone_identifier':
    Invalid identifier: US/Pacific (TZInfo::InvalidTimezoneIdentifier)

This installs the `tzinfo-data` gem to stop using the system zoneinfo.

This fixed my problem, but I also think this is a useful change for
others. Now everyone can rely on the same database instead of their
system's, which may vary between developers.

[0]: https://tzinfo.github.io/