Remove target="_blank" from join.md (#473)
Open the form response (typically the `/welcome` page) in the same tab
instead of a new tab. The current behavior leaves a tab open with the
form and data populated, which caused re-submissions. This change should
reduce the behavior of users coming back later, forgetting whether they
submitted or not, and submitting again.

Ref:
https://www.notion.so/Redirect-from-join-form-Welcome-on-submit-stay-on-same-tab-2b876edfa0728014be18dbaa9efbb103